Split board controller into 3 parts
