+++ src/main/java/com/example/fiction_place1/domain/board/controller/DeleteBoardController.java
@@ -1,0 +1,25 @@
+package com.example.fiction_place1.domain.board.controller;
+
+import com.example.fiction_place1.domain.board.entity.Board;
+import com.example.fiction_place1.domain.board.service.BoardService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+
+@Controller
+@RequiredArgsConstructor
+public class DeleteBoardController {
+    private final BoardService boardService;
+
+    //게시글 삭제
+    @PreAuthorize("isAuthenticated()")
+    @GetMapping("/board/delete/{id}")
+    public String deleteBoard(@PathVariable("id") Long id){
+        Board board = this.boardService.getBoard(id);
+        this.boardService.delete(board);
+        Long boardTypeId = board.getBoardType().getId();  // 게시글의 boardTypeId 가져오기
+        return String.format("redirect:/board?boardTypeId=%d", boardTypeId); // boardTypeId로 리다이렉트
+    }
+}
