--- src/main/resources/templates/job_board.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>채용공고 게시판</title>
-</head>
-<body>
-<h1>채용공고 게시판</h1>
-
-<div>
-    <a href="/board/job/create">게시글 작성</a>
-</div>
-
-<ul>
-    <th:block th:each="board : ${boards}">
-        <li>
-            <h3 th:text="${board.title}"></h3>
-        </li>
-    </th:block>
-</ul>
-</body>
-</html>
